Shared Clipboard/Transfers: Made the context ID 64-bit wide.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@81346 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ include/VBox/VBoxGuestLib.h
@@ -603,7 +603,7 @@ typedef struct VBGLR3SHCLCMDCTX
     /** IN: HGCM client ID to use for communication. */
     uint32_t                  uClientID;
     /** IN/OUT: Context ID to retrieve or to use. */
-    uint32_t                  uContextID;
+    uint64_t                  uContextID;
     /** IN: Protocol version to use. */
     uint32_t                  uProtocolVer;
     /** IN: Protocol flags. Currently unused. */
